feat⭐:  mensagem de error custom

+++ src/storage/group/groupCreate.ts
@@ -1,3 +1,4 @@
+import { AppError } from '@utils/AppError';
 import AsyncStorage from '@react-native-async-storage/async-storage';
 
 import { GROUP_COLLECTION } from '@storage/storageConfig';
@@ -6,6 +7,13 @@ import { groupsGetAll } from './groupsGetAll';
 export async function groupCreate(newGroup: string) {
   try {
     const storedGroups = await groupsGetAll();
+
+    const groupAlreadyExists = storedGroups.includes(newGroup);
+
+    if(groupAlreadyExists) {
+      throw new AppError('Já existe um grupo cadastrado com esse nome.')
+    }
+
     const storage = JSON.stringify([...storedGroups, newGroup])
     await AsyncStorage.setItem(GROUP_COLLECTION, storage);
